# src/main/java/org/openapitools/api/PaymentInitiationApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import com.hiberus.payment.generated.api.model.Error;
import com.hiberus.payment.generated.api.model.PaymentOrder;
import com.hiberus.payment.generated.api.model.PaymentOrderRequest;
import com.hiberus.payment.generated.api.model.PaymentOrderStatus;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.server.ServerWebExchange;
import reactor.core.publisher.Flux;
import reactor.core.publisher.Mono;
import org.springframework.http.codec.multipart.Part;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-11-17T18:46:16.062694-05:00[America/Guayaquil]")
@Validated
@Tag(name = "Payment Orders", description = "the Payment Orders API")
public interface PaymentInitiationApi {

    /**
     * GET /payment-initiation/payment-orders/{id} : Retrieve a payment order
     *
     * @param id  (required)
     * @return Payment order retrieved successfully (status code 200)
     *         or Payment order not found (status code 404)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "getPaymentOrder",
        summary = "Retrieve a payment order",
        tags = { "Payment Orders" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Payment order retrieved successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PaymentOrder.class)),
                @Content(mediaType = "application/problem+json", schema = @Schema(implementation = PaymentOrder.class))
            }),
            @ApiResponse(responseCode = "404", description = "Payment order not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class)),
                @Content(mediaType = "application/problem+json", schema = @Schema(implementation = Error.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class)),
                @Content(mediaType = "application/problem+json", schema = @Schema(implementation = Error.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/payment-initiation/payment-orders/{id}",
        produces = { "application/json", "application/problem+json" }
    )
    
    default Mono<ResponseEntity<PaymentOrder>> getPaymentOrder(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") String id,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        Mono<Void> result = Mono.empty();
        exchange.getResponse().setStatusCode(HttpStatus.NOT_IMPLEMENTED);
        for (MediaType mediaType : exchange.getRequest().getHeaders().getAccept()) {
            if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                String exampleString = "{ \"externalReference\" : \"EXT-1\", \"debtorAccount\" : { \"iban\" : \"EC12DEBTOR\" }, \"requestedExecutionDate\" : \"2025-10-31T00:00:00.000+00:00\", \"remittanceInformation\" : \"Factura 001-123\", \"lastUpdateDate\" : \"2023-11-15T10:30:00Z\", \"creditorAccount\" : { \"iban\" : \"EC12DEBTOR\" }, \"instructedAmount\" : { \"amount\" : 150.75, \"currency\" : \"USD\" }, \"id\" : \"PO-0001\", \"creationDate\" : \"2023-11-15T10:30:00Z\", \"status\" : \"PENDING\" }";
                result = ApiUtil.getExampleResponse(exchange, MediaType.valueOf("application/json"), exampleString);
                break;
            }
            if (mediaType.isCompatibleWith(MediaType.valueOf("application/problem+json"))) {
                String exampleString = "Custom MIME type example not yet supported: application/problem+json";
                result = ApiUtil.getExampleResponse(exchange, MediaType.valueOf(""), exampleString);
                break;
            }
        }
        return result.then(Mono.empty());

    }


    /**
     * GET /payment-initiation/payment-orders/{id}/status : Retrieve the status of a payment order
     *
     * @param id  (required)
     * @return Payment order status retrieved successfully (status code 200)
     *         or Payment order not found (status code 404)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "getPaymentOrderStatus",
        summary = "Retrieve the status of a payment order",
        tags = { "Payment Orders" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Payment order status retrieved successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PaymentOrderStatus.class)),
                @Content(mediaType = "application/problem+json", schema = @Schema(implementation = PaymentOrderStatus.class))
            }),
            @ApiResponse(responseCode = "404", description = "Payment order not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class)),
                @Content(mediaType = "application/problem+json", schema = @Schema(implementation = Error.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class)),
                @Content(mediaType = "application/problem+json", schema = @Schema(implementation = Error.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/payment-initiation/payment-orders/{id}/status",
        produces = { "application/json", "application/problem+json" }
    )
    
    default Mono<ResponseEntity<PaymentOrderStatus>> getPaymentOrderStatus(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") String id,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        Mono<Void> result = Mono.empty();
        exchange.getResponse().setStatusCode(HttpStatus.NOT_IMPLEMENTED);
        for (MediaType mediaType : exchange.getRequest().getHeaders().getAccept()) {
            if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                String exampleString = "{ \"lastUpdateDate\" : \"2023-11-15T10:30:00Z\", \"id\" : \"PO-0001\", \"status\" : \"PENDING\" }";
                result = ApiUtil.getExampleResponse(exchange, MediaType.valueOf("application/json"), exampleString);
                break;
            }
            if (mediaType.isCompatibleWith(MediaType.valueOf("application/problem+json"))) {
                String exampleString = "Custom MIME type example not yet supported: application/problem+json";
                result = ApiUtil.getExampleResponse(exchange, MediaType.valueOf(""), exampleString);
                break;
            }
        }
        return result.then(Mono.empty());

    }


    /**
     * POST /payment-initiation/payment-orders : Initiate a payment order
     *
     * @param paymentOrderRequest  (required)
     * @return Payment order created successfully (status code 201)
     *         or Invalid request (status code 400)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "initiatePaymentOrder",
        summary = "Initiate a payment order",
        tags = { "Payment Orders" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Payment order created successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PaymentOrder.class)),
                @Content(mediaType = "application/problem+json", schema = @Schema(implementation = PaymentOrder.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid request", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class)),
                @Content(mediaType = "application/problem+json", schema = @Schema(implementation = Error.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class)),
                @Content(mediaType = "application/problem+json", schema = @Schema(implementation = Error.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/payment-initiation/payment-orders",
        produces = { "application/json", "application/problem+json" },
        consumes = { "application/json" }
    )
    
    default Mono<ResponseEntity<PaymentOrder>> initiatePaymentOrder(
        @Parameter(name = "PaymentOrderRequest", description = "", required = true) @Valid @RequestBody Mono<PaymentOrderRequest> paymentOrderRequest,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        Mono<Void> result = Mono.empty();
        exchange.getResponse().setStatusCode(HttpStatus.NOT_IMPLEMENTED);
        for (MediaType mediaType : exchange.getRequest().getHeaders().getAccept()) {
            if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                String exampleString = "{ \"externalReference\" : \"EXT-1\", \"debtorAccount\" : { \"iban\" : \"EC12DEBTOR\" }, \"requestedExecutionDate\" : \"2025-10-31T00:00:00.000+00:00\", \"remittanceInformation\" : \"Factura 001-123\", \"lastUpdateDate\" : \"2023-11-15T10:30:00Z\", \"creditorAccount\" : { \"iban\" : \"EC12DEBTOR\" }, \"instructedAmount\" : { \"amount\" : 150.75, \"currency\" : \"USD\" }, \"id\" : \"PO-0001\", \"creationDate\" : \"2023-11-15T10:30:00Z\", \"status\" : \"PENDING\" }";
                result = ApiUtil.getExampleResponse(exchange, MediaType.valueOf("application/json"), exampleString);
                break;
            }
            if (mediaType.isCompatibleWith(MediaType.valueOf("application/problem+json"))) {
                String exampleString = "Custom MIME type example not yet supported: application/problem+json";
                result = ApiUtil.getExampleResponse(exchange, MediaType.valueOf(""), exampleString);
                break;
            }
        }
        return result.then(paymentOrderRequest).then(Mono.empty());

    }

}